Move output comparison code into shared helper function
